requirementsv1.txt
django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
pyfcm==1.5.4
whitenoise==6.6.0
dj-database-url==2.1.0
django-cors-headers==4.3.1
django-filter==23.5
django-celery-beat==2.5.0
django-redis==5.2.0
drf-yasg==1.21.7
django-admin-material-dashboard==1.0.22